Sketch out support for OAuth handlers.
Moved would-be handlers in the example to just be default supported,
like the OAuth login and authorize endpoints. Restructured the Context
so it can clarify between a user resource path and a request bearing the
authorization of a user. Refactored the API for an AP or AP+Web call to
be more sane.

Still some todos. But overall should increase the value of the framework
and save boilerplate for apcore users.