Introduce model & service support for NodeInfo.
This includes anonymizing functionality around total users and user
activity numbers for small instances, to ensure that folks that would
scrape this data are not able to attribute changes in counts of active
users to specific users.

A simple anonymizing statistic of uncertainty around the mean using
rejection sampling ensures that the metrics make sense and are
approximately correct for small values, but precisely correct for
instances larger in size.