Add methods to set the server's instance profile.
This will allow such data to be reflected in the NodeInfo fetches.
However, this requires setting metadata on the server instance actor
that does not exist, so we're potentially creating a new ActivityStreams
vocabulary.

The work to craft the SQL queries is still TBD, as well as the
associated test model database binary.