Framework provides Send Accept/Reject Follow
Updating the example application to manage followers, I believe that the
ActivityStreams logic would probably all be shared by client
applications. The example handler and apcore framework changes are here,
but still missing are the web pages to allow the user to actually see
what is going on and trigger behaviors.